+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [#121](https://github.com/deviceinsight/kafkactl/issues/121) Support message consumption with a specified timestamp range
+
 
 ## 3.0.3 - 2023-02-01
 ### Changed